backend/foodgram/requirements.txt
Django==3.2.3
django-filter==23.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-cors-headers==3.13.0
djoser==2.1.0
gunicorn==20.0.4
psycopg2-binary==2.9.3
Pillow==9.0.0
python-dotenv==1.0.0